requirements.txt
asgiref==3.2.3
Django==3.0.4
django-autocomplete-light==3.5.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3-datetimepicker==2.2.3
django-crispy-forms==1.9.0
django-tempus-dominus==5.1.2.13
extras==1.0.0
psycopg2-binary==2.8.5
pytz==2019.3
six==1.15.0
sqlparse==0.3.1
